requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Redis (async support needed for infra WebSocket/pubsub layer)
redis[asyncio]>=5.0.0

# AWS S3
boto3>=1.34.0

# Environment variables
python-dotenv>=1.0.0

# Data validation (used by FastAPI models)
pydantic>=2.0.0

# Testing
pytest>=8.0.0
httpx>=0.27.0          # required by FastAPI TestClient
